add: ログインユーザーの作成したアイテムだけを更新・削除できるようトークン判定処理を追加

+++ app/api/item/delete/[id]/route.js
@@ -3,10 +3,16 @@ import { ItemModel } from "@/app/utils/schemaModels";
 import { NextResponse } from "next/server";
 
 export async function DELETE(request, context) {
+  const reqBody = await request.json();
   try {
     await connectDB();
-    await ItemModel.deleteOne({ _id: context.params.id });
-    return NextResponse.json({ message: "アイテム削除" });
+    const singleItem = await ItemModel.findById(context.params.id);
+    if (singleItem.email === reqBody.email) {
+      await ItemModel.deleteOne({ _id: context.params.id });
+      return NextResponse.json({ message: "アイテム削除成功" });
+    } else {
+      return NextResponse.json({ message: "他の人が作成したアイテムです" });
+    }
   } catch (error) {
     return NextResponse.json({ message: "エラー" });
   }
